fix: address second-round Copilot review findings
- Clear debounce timer on Enter to prevent duplicate IPC calls
- Move getCategoryColor to shared lib/error-categories.ts (removes
  cross-feature coupling between dialogs and log-view)
- Error code spans: add role="button", tabIndex, keyboard handlers
  for Enter/Space; only stopPropagation when onSpanClick is defined
- Use UTF-16 code unit counting for span offsets (chars().count()
  gives wrong indices after emoji/non-BMP characters in JS)
- Add emoji prefix regression test for UTF-16 correctness
- Treat pure numeric input (e.g., "5", "1603") as code pattern
  for instant lookup instead of 300ms debounce
